LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-attribute-changes-expected.txt


FAIL Meta color-scheme initially 'dark'. assert_equals: expected "dark" but got "normal"
PASS Removed name attribute from meta color-scheme.
FAIL Set meta name to color-scheme. assert_equals: expected "dark" but got "normal"
PASS Set content attribute of meta color-scheme to empty string.
PASS Set content attribute of meta color-scheme to an invalid value.
FAIL Set content attribute of meta color-scheme to 'light'. assert_equals: expected "light" but got "normal"
PASS Removed the content attribute of meta color-scheme.
